This patch makes sure that specifying an initrd in the LAB config file is actually respected (that was a bug).
It also moves the LAB config file to /lab.conf on the first partition. The
kexec binary is now also expected to live in /kexec on the first partition. 

The reason for putting the LAB config file and kexec binary in / on partition 1
rather than /etc and /sbin respectively, is that often (at least in our setups)
/boot is the first partition, and is a simple software raid-1 across however
many disks there are in the system. The rest of the partitions can be software
raid-5, LVM on sw raid, etc. We do things this way because GRUB does not
understand software raid nor LVM.

Not needing MD or LVM drivers in the kernel that lives in rom makes it smaller,
which is desirable. Putting a static kexec in / on the first partition means
that it will not be managed by a package manager, which could overwrite it with
a dynamic version one day, which would break the next boot...

Signed-off-by: Ward Vandewege <ward@gnu.org>
Acked-by: Jordan Crouse <jordan.crouse@amd.com>