This patch applies to the Supermicro H8DMR board.
It upgrades the linuxbios version used to 2996 so that we can use a VGA option rom.

It also fixes the devices FILO looks on to find the grub config file to hde1,
which is what the first drive is recognized as on an MCP55-based board like this one.

This is a trivial patch.

Signed-off-by: Ward Vandewege <ward@gnu.org>
Acked-by: Ward Vandewege <ward@gnu.org>