This patch adds extract targets, as well as some config targets: busybox-config, uclibc-config, kernel-config and filo-config.
The extract targets will just extract the component(s) under the work
subdirectory.

The config targets will run a configure command for the component, and then
copy the resulting custom config file to the conf directory for the package
using a naming convention that indicates the payload, vendor, board and (for
the kernel and uclibc) architecture.

That config file will then be used to build the image when 'make' is issued,
overriding the default buildrom config for that package/payload - but only if
there is a match for payload, vendor and board (and for uclibc and the kernel,
architecture).

The config targets are somewhat smart: they will copy an existing custom config
file back to the source directory so that doing a make something-config when a
custom config file exists will have the effect of editing the custom config,
rather than overwriting it.

If a custom config file is used, buildrom will say so on stdout.

Signed-off-by: Ward Vandewege <ward@gnu.org>
Acked-by: Myles Watson <myles@pel.cs.byu.edu>