This patch adds an optional delay before the LAB environment kexecs the on-disk kernel. This allows entering the busybox shell, which can be useful for maintenance and debugging. The delay is configurable from kconfig, and defaults to 5 seconds.
The patch enables CONFIG_ASH_READ_TIMEOUT in the uclibc configuration because
the 'read' command in the ash shell does not support the '-t' timeout parameter
otherwise.

It also changes the default for including a statically compiled copy of kexec
into the LAB payload to no, because this makes the payload too large for 1MB
rom chips.

Tested on real hardware.

Signed-off-by: Ward Vandewege <ward@gnu.org>
Acked-by: Jordan Crouse <jordan.crouse@amd.com>