[BUILDROM] Fixup VSA fetching code
In preparation for v3 and the upcoming openvsa changes,
rearrange the VSA fetching code.  We now fetch an 
uncompressed blob, and do the compressing within buildrom.
I'm mainly concentrating on v2 right now (this is not quite
v3 friendly), but it will be.

Also, work around a hopefully temporary bug on the AMD
servers.

Signed-off-by: Jordan Crouse <jordan.crouse@amd.com>
Acked-by: Ward Vandewege <ward@gnu.org>