Move from our grub2 fork to mainline grub2, now that the features we need have been included there.
This patch also makes sure that LZMA compression of the payload is
automatically disabled for filo, etherboot and grub2, since it breaks booting
in those cases.

It also removes a few bash-isms in fetchsvn.sh.

Many, many thanks to Robert Milan for fixing the bugs in grub2 that I found
while testing.

This was build- and boot-tested on qemu, on v2 and v3.

Signed-off-by: Ward Vandewege <ward@gnu.org>
Acked-by: Jordan Crouse <jordan@cosmicpenguin.net>